// src/interfaces/ProjectsWebInfo.ts
export interface ProjectsWebInfo {
    dateCreated: string;
    description: string;
    githublink: string;
    id: number;
    image: string[];
    link: string;
    name: string;
    notes: string;
    previewLink: string;
    sequence: number;
    tags: string;
    technology: string;
    type: string;
    videoAds: string;
    videoDemo: string;
}